+++ packages/common/src/types/field.ts
@@ -1,4 +1,4 @@
-import type { AdditionalMetric, currencies } from '..';
+import type { AdditionalMetric, currencies, DefaultTimeDimension } from '..';
 import { CompileError } from './errors';
 import { type MetricFilterRule } from './filter';
 import { type TimeFrames } from './timeFrames';
@@ -445,6 +445,7 @@ export interface Metric extends Field {
     formatOptions?: CustomFormat;
     dimensionReference?: string; // field id of the dimension this metric is based on
     requiredAttributes?: Record<string, string | string[]>; // Required attributes for the dimension this metric is based on
+    defaultTimeDimension?: DefaultTimeDimension; // Default time dimension for the metric when the user has not specified a time dimension
 }
 
 export const isFilterableDimension = (